+++ slo-workload/src/create.ts
@@ -1,45 +1,37 @@
-import {
-  Column,
-  Driver,
-  ExecuteQuerySettings,
-  PartitioningPolicy,
-  TableDescription,
-  TableProfile,
-  TypedData,
-  Types,
-} from 'ydb-sdk'
+import { Column, Driver, ExecuteQuerySettings, TableDescription, Types, Ydb } from 'ydb-sdk'
 
-import { PackGenerator } from './utils/PackGenerator'
-import { StructValue } from './utils/StructValue'
 import {
   TABLE_NAME,
-  TABLE_PARTITION_COUNT,
+  TABLE_MIN_PARTITION_COUNT,
+  TABLE_MAX_PARTITION_COUNT,
   GENERATOR_DATA_COUNT,
-  GENERATOR_PACK_SIZE,
+  TABLE_PARTITION_SIZE,
 } from './utils/defaults'
+import { QueryBuilder } from './utils/QueryBuilder'
+import { DataGenerator } from './utils/DataGenerator'
 
 export async function create(
   driver: Driver,
-  db: string,
   tableName?: string,
-  _partitionsCount?: string,
+  _minPartitionsCount?: string,
+  _maxPartitionsCount?: string,
+  _partitionSize?: string,
   _dataCount?: string
 ) {
   if (!tableName) tableName = TABLE_NAME
 
-  let partitionsCount: number = TABLE_PARTITION_COUNT
-  if (_partitionsCount) {
-    partitionsCount = parseIntNumber(_partitionsCount)
-    partitionsCount = Math.floor(partitionsCount)
-  }
+  let minPartitionsCount: number = TABLE_MIN_PARTITION_COUNT
+  let maxPartitionsCount: number = TABLE_MAX_PARTITION_COUNT
+  let partitionSize: number = TABLE_PARTITION_SIZE
+
+  if (_minPartitionsCount) minPartitionsCount = parseIntNumber(_minPartitionsCount)
+  if (_maxPartitionsCount) maxPartitionsCount = parseIntNumber(_maxPartitionsCount)
+  if (_partitionSize) partitionSize = parseIntNumber(_partitionSize)
 
   let dataCount: number = GENERATOR_DATA_COUNT
-  if (_dataCount) {
-    dataCount = parseIntNumber(_dataCount)
-    dataCount = Math.floor(dataCount)
-  }
+  if (_dataCount) dataCount = parseIntNumber(_dataCount)
 
-  await createTable(driver, tableName, partitionsCount)
+  await createTable(driver, tableName, minPartitionsCount, maxPartitionsCount, partitionSize)
   await generateInitialContent(driver, tableName, dataCount)
 
   process.exit(0)
@@ -48,22 +40,40 @@ export async function create(
 function parseIntNumber(numberCandidate: string): number {
   const n = Number(numberCandidate)
   if (isNaN(n)) throw Error('Not a number')
-  return n
+  return Math.floor(n)
 }
 
-async function createTable(driver: Driver, tableName: string, partitions: number) {
-  console.log('Create table', { task: 'createTable', tableName, partitions })
+async function createTable(
+  driver: Driver,
+  tableName: string,
+  minPartitionsCount: number,
+  maxPartitionsCount: number,
+  partitionSize: number
+) {
+  console.log('Create table', {
+    task: 'createTable',
+    tableName,
+    minPartitionsCount,
+    maxPartitionsCount,
+    partitionSize,
+  })
+
   await driver.tableClient.withSession(async (session) => {
-    const profile = new TableProfile().withPartitioningPolicy(
-      new PartitioningPolicy().withUniformPartitions(partitions)
-    )
     const tableDescription = new TableDescription()
-      .withColumn(new Column('object_id_key', Types.optional(Types.UINT32)))
-      .withColumn(new Column('object_id', Types.optional(Types.UINT32)))
-      .withColumn(new Column('timestamp', Types.optional(Types.UINT64)))
-      .withColumn(new Column('payload', Types.optional(Types.UTF8)))
-      .withPrimaryKeys('object_id_key', 'object_id')
-      .withProfile(profile)
+      .withColumn(new Column('hash', Types.optional(Types.UINT64)))
+      .withColumn(new Column('id', Types.optional(Types.UINT64)))
+      .withColumn(new Column('payload_str', Types.optional(Types.UTF8)))
+      .withColumn(new Column('payload_double', Types.optional(Types.DOUBLE)))
+      .withColumn(new Column('payload_timestamp', Types.optional(Types.TIMESTAMP)))
+      .withColumn(new Column('payload_hash', Types.optional(Types.UINT64)))
+      .withPrimaryKeys('hash', 'id')
+
+    tableDescription.partitioningSettings = {
+      minPartitionsCount: minPartitionsCount,
+      maxPartitionsCount: maxPartitionsCount,
+      partitionSizeMb: partitionSize,
+      partitioningBySize: Ydb.FeatureFlag.Status.ENABLED,
+    }
 
     await session.createTable(tableName, tableDescription)
     console.log('Table created')
@@ -72,30 +82,23 @@ async function createTable(driver: Driver, tableName: string, partitions: number
 
 async function generateInitialContent(driver: Driver, tableName: string, count: number) {
   console.log('Generate initial content', { task: 'generateInitialContent', tableName, count })
-  const generator = new PackGenerator(count, GENERATOR_PACK_SIZE, 0)
 
-  const query = `--!syntax_v1
-    DECLARE $items AS
-    List<Struct<
-    object_id_key: Uint32,
-    object_id: Uint32,
-    timestamp: Uint64,
-    payload: Utf8>>;
-    UPSERT INTO \`${tableName}\`
-    SELECT * FROM AS_TABLE($items);`
+  const count5p = Math.floor(count / 20) // every 5% of count for logs
 
-  let batch: StructValue[]
-  while ((batch = generator.get()).length > 0) {
-    // TODO: add executor
-    await driver.tableClient.withSession(async (session) => {
-      await session.executeQuery(
-        query,
-        { $items: TypedData.asTypedCollection(batch) },
-        { commitTx: true, beginTx: { serializableReadWrite: {} } },
-        new ExecuteQuerySettings().withKeepInCache(true)
-      )
-    }, 10000)
-    console.log(`Successfully inserted ${batch.length} rows`)
-  }
+  const qb = new QueryBuilder(tableName, 10, 10) // stub timeouts
+  DataGenerator.setMaxId(0)
+
+  await Promise.all(
+    Array.apply(null, Array(count)).map(() => {
+      return driver.tableClient.withSession(async (session) => {
+        await session.executeQuery(
+          qb.writeQuery,
+          { ...DataGenerator.getUpsertData() },
+          { commitTx: true, beginTx: { serializableReadWrite: {} } },
+          new ExecuteQuerySettings().withKeepInCache(true)
+        )
+      }, 10000)
+    })
+  )
   console.log('Initial content generated')
 }