+++ slo-workload/src/metricsJob.ts
@@ -5,16 +5,13 @@ export class MetricsJob {
   private endTime: number
   private promise: Promise<void>
 
-  constructor(
-    private executor: Executor,
-    private reportPeriod: number = 1000,
-    private time: number
-  ) {
-    this.endTime = new Date().valueOf() + time * 1000
+  constructor(private executor: Executor, private reportPeriod: number = 1000, endTime: number) {
+    this.endTime = new Date().valueOf() + endTime * 1000
 
     this.promise = new Promise((resolve) => {
       this.intervalId = setInterval(() => {
         if (new Date().valueOf() > this.endTime) {
+          console.log('Metrics job ended')
           clearInterval(this.intervalId)
           return resolve()
         }