+++ slo-workload/src/utils/defaults.ts
@@ -1,8 +1,9 @@
 export const TABLE_NAME = 'slo-test'
-export const TABLE_PARTITION_COUNT = 64
+export const TABLE_MIN_PARTITION_COUNT = 6
+export const TABLE_MAX_PARTITION_COUNT = 1000
+export const TABLE_PARTITION_SIZE = 1
 
 export const GENERATOR_DATA_COUNT = 1000
-export const GENERATOR_PACK_SIZE = 100
 
 export const READ_RPS = 1000
 export const READ_TIMEOUT = 70 // milliseconds